+++ packages/website/demos/AxesHelper/index.tsx
@@ -3,7 +3,7 @@ import Button from '@material-ui/core/Button'
 import { Canvas } from '@react-vertex/core'
 import { useMeasure } from 'react-use'
 import { convertHex } from '@react-vertex/color-hooks'
-import DemoWrapper from '../DemoWrapper'
+import { DemoWrapper } from '../../components/DemoWrapper'
 import { Scene } from './Scene'
 
 const clearColor = convertHex('#323334')
