+++ packages/website/demos/FlockingBirds/index.tsx
@@ -4,7 +4,7 @@ import Typography from '@material-ui/core/Typography'
 import { Canvas } from '@react-vertex/core'
 import { useMeasure } from 'react-use'
 import { convertHex } from '@react-vertex/color-hooks'
-import DemoWrapper from '../DemoWrapper'
+import { DemoWrapper } from '../../components/DemoWrapper'
 import { Scene } from './Scene'
 
 const clearColor = convertHex('#323334')
