--- packages/website/demos/FluidSimulation/README.md
@@ -1,4 +1,0 @@
-## Fluid Simulation
-
-A recreation of the awesome demo at:
-https://github.com/PavelDoGreat/WebGL-Fluid-Simulation
