--- packages/website/demos/FluidSimulation/config.ts
@@ -1,12 +1,0 @@
-export const SIM_RESOLUTION = 128
-export const DYE_RESOLUTION = 512
-export const DENSITY_DISSIPATION = 0.97
-export const VELOCITY_DISSIPATION = 0.98
-export const PRESSURE_DISSIPATION = 0.8
-export const PRESSURE_ITERATIONS = 20
-export const CURL = 30
-export const SPLAT_RADIUS = 0.5
-export const SHADING = true
-export const COLORFUL = true
-export const BACK_COLOR = { r: 0, g: 255, b: 0 }
-export const TRANSPARENT = false
