--- packages/website/demos/FluidSimulation/customHooks/index.ts
@@ -1,5 +1,0 @@
-export { default as usePrograms } from './usePrograms'
-export { default as useFramebuffers } from './useFramebuffers'
-export { default as usePointers } from './usePointers'
-export { default as useFormats } from './useFormats'
-export { default as useResolution } from './useResolution'
