--- packages/website/demos/FluidSimulation/customHooks/usePrograms/shaders/advection.frag.ts
@@ -1,20 +1,0 @@
-export default `
-  precision highp float;
-  precision highp sampler2D;
-
-  // This shader is from the awesome demo at:
-  // https://github.com/PavelDoGreat/WebGL-Fluid-Simulation
-
-  varying vec2 vUv;
-  uniform sampler2D uVelocity;
-  uniform sampler2D uSource;
-  uniform vec2 texelSize;
-  uniform float dt;
-  uniform float dissipation;
-
-  void main () {
-    vec2 coord = vUv - dt * texture2D(uVelocity, vUv).xy * texelSize;
-    gl_FragColor = dissipation * texture2D(uSource, coord);
-    gl_FragColor.a = 1.0;
-  }
-`
