--- packages/website/demos/FluidSimulation/customHooks/usePrograms/shaders/advection.manual.frag.ts
@@ -1,36 +1,0 @@
-export default `
-  precision highp float;
-  precision highp sampler2D;
-
-  // This shader is from the awesome demo at:
-  // https://github.com/PavelDoGreat/WebGL-Fluid-Simulation
-
-  varying vec2 vUv;
-  uniform sampler2D uVelocity;
-  uniform sampler2D uSource;
-  uniform vec2 texelSize;
-  uniform vec2 dyeTexelSize;
-  uniform float dt;
-  uniform float dissipation;
-
-  vec4 bilerp (sampler2D sam, vec2 uv, vec2 tsize) {
-    vec2 st = uv / tsize - 0.5;
-
-    vec2 iuv = floor(st);
-    vec2 fuv = fract(st);
-
-    vec4 a = texture2D(sam, (iuv + vec2(0.5, 0.5)) * tsize);
-    vec4 b = texture2D(sam, (iuv + vec2(1.5, 0.5)) * tsize);
-    vec4 c = texture2D(sam, (iuv + vec2(0.5, 1.5)) * tsize);
-    vec4 d = texture2D(sam, (iuv + vec2(1.5, 1.5)) * tsize);
-
-    return mix(mix(a, b, fuv.x), mix(c, d, fuv.x), fuv.y);
-  }
-
-  void main () {
-    vec2 coord = vUv - dt * bilerp(uVelocity, vUv, texelSize).xy * texelSize;
-
-    gl_FragColor = dissipation * bilerp(uSource, coord, dyeTexelSize);
-    gl_FragColor.a = 1.0;
-  }
-`
