--- packages/website/demos/FluidSimulation/customHooks/usePrograms/shaders/index.ts
@@ -1,13 +1,0 @@
-export { default as vert } from './base.vert'
-export { default as advection } from './advection.frag'
-export { default as advectionManual } from './advection.manual.frag'
-export { default as background } from './background.frag'
-export { default as clear } from './clear.frag'
-export { default as color } from './color.frag'
-export { default as curl } from './curl.frag'
-export { default as displayShading } from './display.shading.frag'
-export { default as divergence } from './divergence.frag'
-export { default as gradient } from './gradient.subtract.frag'
-export { default as pressure } from './pressure.frag'
-export { default as splat } from './splat.frag'
-export { default as vorticity } from './vorticity.frag'
