--- packages/website/demos/FluidSimulation/customHooks/usePrograms/shaders/pressure.frag.ts
@@ -1,33 +1,0 @@
-export default `
-  precision mediump float;
-  precision mediump sampler2D;
-
-  // This shader is from the awesome demo at:
-  // https://github.com/PavelDoGreat/WebGL-Fluid-Simulation
-
-  varying highp vec2 vUv;
-  varying highp vec2 vL;
-  varying highp vec2 vR;
-  varying highp vec2 vT;
-  varying highp vec2 vB;
-  uniform sampler2D uPressure;
-  uniform sampler2D uDivergence;
-
-  vec2 boundary (vec2 uv) {
-    return uv;
-    // uncomment if you use wrap or repeat texture mode
-    // uv = min(max(uv, 0.0), 1.0);
-    // return uv;
-  }
-
-  void main () {
-    float L = texture2D(uPressure, boundary(vL)).x;
-    float R = texture2D(uPressure, boundary(vR)).x;
-    float T = texture2D(uPressure, boundary(vT)).x;
-    float B = texture2D(uPressure, boundary(vB)).x;
-    float C = texture2D(uPressure, vUv).x;
-    float divergence = texture2D(uDivergence, vUv).x;
-    float pressure = (L + R + B + T - divergence) * 0.25;
-    gl_FragColor = vec4(pressure, 0.0, 0.0, 1.0);
-  }
-`
