--- packages/website/demos/FluidSimulation/customHooks/usePrograms/shaders/splat.frag.ts
@@ -1,23 +1,0 @@
-export default `
-  precision highp float;
-  precision highp sampler2D;
-
-  // This shader is from the awesome demo at:
-  // https://github.com/PavelDoGreat/WebGL-Fluid-Simulation
-
-  varying vec2 vUv;
-
-  uniform sampler2D uTarget;
-  uniform float aspectRatio;
-  uniform vec3 color;
-  uniform vec2 point;
-  uniform float radius;
-
-  void main () {
-    vec2 p = vUv - point.xy;
-    p.x *= aspectRatio;
-    vec3 splat = exp(-dot(p, p) / radius) * color;
-    vec3 base = texture2D(uTarget, vUv).xyz;
-    gl_FragColor = vec4(base + splat, 1.0);
-  }
-`
