--- packages/website/demos/FluidSimulation/customHooks/usePrograms/shaders/vorticity.frag.ts
@@ -1,33 +1,0 @@
-export default `
-  precision highp float;
-  precision highp sampler2D;
-
-  // This shader is from the awesome demo at:
-  // https://github.com/PavelDoGreat/WebGL-Fluid-Simulation
-
-  varying vec2 vUv;
-  varying vec2 vL;
-  varying vec2 vR;
-  varying vec2 vT;
-  varying vec2 vB;
-  uniform sampler2D uVelocity;
-  uniform sampler2D uCurl;
-  uniform float curl;
-  uniform float dt;
-
-  void main () {
-    float L = texture2D(uCurl, vL).x;
-    float R = texture2D(uCurl, vR).x;
-    float T = texture2D(uCurl, vT).x;
-    float B = texture2D(uCurl, vB).x;
-    float C = texture2D(uCurl, vUv).x;
-
-    vec2 force = 0.5 * vec2(abs(T) - abs(B), abs(R) - abs(L));
-    force /= length(force) + 0.0001;
-    force *= curl * C;
-    force.y *= -1.0;
-
-    vec2 vel = texture2D(uVelocity, vUv).xy;
-    gl_FragColor = vec4(vel + force * dt, 0.0, 1.0);
-  }
-`
