--- packages/website/pages/demo-fluid-simulation.tsx
@@ -1,3 +1,0 @@
-import FluidSimulation from '../demos/FluidSimulation'
-
-export default FluidSimulation
